Address changes mentioned in a discussion
